+++ Mmc.MonoGame.Noise/src/Transforms/ScaleNoise.cs
@@ -4,18 +4,18 @@
     {
         private INoise SourceNoise { get; init; }
 
-        private float XScale { get; init; }
+        private double XScale { get; init; }
 
-        private float YScale { get; init; }
+        private double YScale { get; init; }
 
-        public ScaleNoise(INoise sourceNoise, float xScale, float yScale)
+        public ScaleNoise(INoise sourceNoise, double xScale, double yScale)
         {
             SourceNoise = sourceNoise;
             XScale = xScale;
             YScale = yScale;
         }
 
-        public float GetValue(float x, float y)
+        public double GetValue(double x, double y)
         {
             return SourceNoise.GetValue(x * XScale, y * YScale);
         }
